Move quad analyzer class to our package
